requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
blis==0.7.11
catalogue==2.0.10
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cymem==2.0.8
demoji==1.1.0
Django==4.2.6
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl#sha256=6215d71a3212690e9aec49408a27e3fe6ad7cd6c715476e93d70dc784041e93e
frozenlist==1.4.0
idna==3.4
Jinja2==3.1.2
langcodes==3.3.0
lxml==4.9.3
MarkupSafe==2.1.3
mock==5.1.0
multidict==6.0.4
murmurhash==1.0.10
numpy==1.26.2
openai==0.28.1
packaging==23.2
pandas==2.1.3
plotly==5.11.0
preshed==3.0.9
pydantic==2.5.2
pydantic_core==2.14.5
python-dateutil==2.8.2
python-docx==1.0.1
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
scikit-learn==1.3.2
six==1.16.0
smart-open==6.4.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.4
srsly==2.4.8
thinc==8.2.1
tomli==2.0.1
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wasabi==1.1.2
weasel==0.3.4
xmltodict==0.13.0
yarl==1.9.2
